backend/requirements.txt
python-binance==1.0.19
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
